requirements.txt
alembic==0.9.1
aniso8601==7.0.0
appdirs==1.4.3
asn1crypto==0.24.0
blinker==1.4
boddle==0.2.8
boto==2.49.0
boto3==1.9.67
botocore==1.12.67
bottle==0.12.13
bottle-mysql==0.3.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==1.9
django-colorfield==0.1.15
django-crispy-forms==1.7.2
django-extensions==2.1.7
django-storages==1.7.1
django-tastypie==0.14.2
djangorestframework-jwt==1.11.0
docutils==0.14
edx-auth-backends==1.2.1
Flask==0.12
Flask-Cors==3.0.3
Flask-HTTPAuth==3.2.4
Flask-Login==0.4.0
Flask-Migrate==2.0.3
Flask-RESTful==0.3.7
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
future==0.17.1
http-basic-auth==1.2.0
idna==2.7
itsdangerous==0.24
Jinja2==2.9.5
jmespath==0.9.3
Mako==1.0.6
MarkupSafe==1.0
mysql-connector-python==8.0.12
oauthlib==2.1.0
packaging==16.8
passlib==1.7.1
pbr==5.3.1
pipenv==2018.11.26
protobuf==3.6.1
psycopg2==2.7.5
pycparser==2.19
pycryptodomex==3.7.2
pyjwkest==1.4.0
PyJWT==1.6.4
PyMySQL==0.9.2
pyparsing==2.2.0
python-dateutil==2.6.1
python-editor==1.0.3
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2018.7
raven==6.10.0
requests==2.20.1
requests-oauthlib==1.0.0
s3transfer==0.1.13
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.7.0
SQLAlchemy==1.1.6
stevedore==1.30.1
urllib3==1.24.1
virtualenv==16.7.2
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.5
Werkzeug==0.12
WTForms==2.2.1
